Finish UI changes for Snake and add the brute force solving method.
This patch added some finishing touches to the base game of Snake.
-Changed UI color scheme of snake, food, and background
-Made snake render as one shape with no gaps between body segments, but
maintain a gap between body lengths when they are side by side.
-made the initial starting position random

Added the first solving Algo for Snake
-Using a brute force approach, this algo searches every grid point on
the board in a repeating track. The Snake will never run into itself,
but the approach is extreamly slow and has the heighest cost in time to
complete the game.

* Snake/snake.py: Modified.
* process-pid-analysis.py: Removed becasue it was a simple script that
  does not match the theme of this repository.