Fixed two major bugs in snake.py
Bug 1:
Food.randomize_position() was not optimal as it would retry new food
positions until it found one that snake was not occupying. If this was
done when snake filled the whole screen then it created an infinate
loop. Fixed this by rewiting the randomize_position() function to take a
list of positions from snake, compare that against our generated list of
board positions, and create a new list of valid positions to choose
randomly from. This greatly reduced the complexity of this function.

Bug 2:
I did not want the game to end by having the snake enevitably fill the
whole screen. I simply added a max_length value to snake so it cant grow
past that size. This is hard coded though and I think should be
generated as a percentage calculation of available board positions from
the start.

Minor fix to how the GRID is calculated in generateGridCoordinates()

*Snake/snake.py: Modified