Finished the Bredth First Search implementation.
snake.py now has the flag --bfs, which will launch that game and use a
bredth first search algorythem to guide the snake to the food. This
method is good bu not perfect. The snake will inevitably get stuck where
it cannot route to the food. When this happend the game will pause with
an error demontrating the inability to route, then reset itself.

I find that the bfs method is a good way to show simple path finding and
its shortcomings in circumstances like this.

*Snake/BFS-testing: Deleted
*Snake/snake.py: Modified