Added the start of A* search.
This patch adds A* search, but I do not have the algorithm quite right
yet. it behaves more like greedy best first seach right now, though its
paths are more clean and direct. I will try to tweak the heuristics and
future patches.

Next I will need to tackle making the snake not trap itself, or to have
the snake atempt to survive until a viable path is found to the food.

*Snake/snake.py: Modified.